+++ library/src/main/java/com/github/paolorotolo/appintro/AppIntroBase.java
@@ -53,6 +53,7 @@ public abstract class AppIntroBase extends AppCompatActivity {
     protected int unselectedIndicatorColor = DEFAULT_COLOR;
     protected View nextButton;
     protected View doneButton;
+    protected View skipButton;
     protected int savedCurrentItem;
     protected ArrayList<PermissionObject> permissionsArray = new ArrayList<>();
 
@@ -65,6 +66,7 @@ public abstract class AppIntroBase extends AppCompatActivity {
     private boolean isImmersiveModeEnabled = false;
     private boolean isImmersiveModeSticky = false;
     private boolean areColorTransitionsEnabled = false;
+    protected boolean skipButtonEnabled = true;
 
     private int currentlySelectedItem = -1;
 
@@ -81,6 +83,7 @@ public abstract class AppIntroBase extends AppCompatActivity {
 
         nextButton = findViewById(R.id.next);
         doneButton = findViewById(R.id.done);
+        skipButton = findViewById(R.id.skip);
 
         mVibrator = (Vibrator) this.getSystemService(VIBRATOR_SERVICE);
         mPagerAdapter = new PagerAdapter(getSupportFragmentManager(), fragments);
@@ -97,6 +100,19 @@ public abstract class AppIntroBase extends AppCompatActivity {
                     mVibrator.vibrate(vibrateIntensity);
                 }
                 onDonePressed(mPagerAdapter.getItem(pager.getCurrentItem()));
+            }
+        });
+
+        skipButton.setOnClickListener(new View.OnClickListener()
+        {
+            @Override
+            public void onClick(@NonNull View v)
+            {
+                if (isVibrateOn)
+                {
+                    mVibrator.vibrate(vibrateIntensity);
+                }
+                onSkipPressed(mPagerAdapter.getItem(pager.getCurrentItem()));
             }
         });
 
@@ -167,6 +183,7 @@ public abstract class AppIntroBase extends AppCompatActivity {
         outState.putBoolean("progressButtonEnabled", progressButtonEnabled);
         outState.putBoolean("nextEnabled", pager.isPagingEnabled());
         outState.putBoolean("nextPagingEnabled", pager.isNextPagingEnabled());
+        outState.putBoolean("skipButtonEnabled", skipButtonEnabled);
         outState.putInt("lockPage", pager.getLockPage());
         outState.putInt("currentItem", pager.getCurrentItem());
 
@@ -182,6 +199,7 @@ public abstract class AppIntroBase extends AppCompatActivity {
 
         this.baseProgressButtonEnabled = savedInstanceState.getBoolean("baseProgressButtonEnabled");
         this.progressButtonEnabled = savedInstanceState.getBoolean("progressButtonEnabled");
+        this.skipButtonEnabled = savedInstanceState.getBoolean("skipButtonEnabled");
         this.savedCurrentItem = savedInstanceState.getInt("currentItem");
         pager.setPagingEnabled(savedInstanceState.getBoolean("nextEnabled"));
         pager.setNextPagingEnabled(savedInstanceState.getBoolean("nextPagingEnabled"));
@@ -233,7 +251,19 @@ public abstract class AppIntroBase extends AppCompatActivity {
      * @param position Position of the new selected slide
      */
     protected void onPageSelected(int position) {
+        setButtonState(skipButton, skipButtonEnabled);
+    }
 
+    public boolean isSkipButtonEnabled() {
+        return skipButtonEnabled;
+    }
+
+    /**
+     * Called when the user clicked the skip button
+     * @param currentFragment Instance of the currently displayed fragment
+     */
+    public void onSkipPressed(Fragment currentFragment) {
+        onSkipPressed();
     }
 
     protected void setScrollDurationFactor(int factor) {
@@ -332,6 +362,14 @@ public abstract class AppIntroBase extends AppCompatActivity {
      * @deprecated Override {@link #onDonePressed(Fragment)} instead
      */
     public void onDonePressed() {
+
+    }
+
+    /**
+     * Called when the user clicked the skip button
+     * @deprecated Override {@link #onSkipPressed(Fragment)} instead
+     */
+    public void onSkipPressed() {
 
     }
